First kick at request handlers in Endpoint.
This is a first attempt at responding to SYN_STREAM frames in
the Endpoint module, using a pluggable request handler.

A request handler has the type

  HeaderBlock
  -> Maybe (IO StreamContent)
  -> IO (HeaderBlock, Maybe ((StreamContent -> IO ()) -> IO ()))

which is a bit on the ugly side. You can think of it as

  HeaderBlock
  -> Maybe Puller
  -> IO (HeaderBlock, Maybe (Pusher -> IO ()))

where

  Puller = IO StreamContent

pulls the next piece of incoming content, and

  Pusher = StreamContent -> IO ()

pushes a piece of outgoing content.

The idea is that given the headers from a SYN_STREAM frame and
an optional puller, the request handler produces a set of initial
response headers for the SYN_REPLY frame, and an optional function
that pushes all the remaining content for the response.

The optional puller is Nothing if and only if the remote endpoint
initiated the stream half-closed. The optional push loop is
Nothing if and only if there is no response other than the initial
response headers.

The SYN_STREAM handler creates a new stream with 'addStream',
and then forks a thread that invokes the request handler. The
thread sends a SYN_REPLY using the initial response headers and,
if the push loop is provided, runs it with the stream pusher
to push the rest of the response. When the response is complete,
the thread removes the stream from the enclosing connection.

The default request handler responds with an HTTP 1.1 404 error
to all requests, carried in the initial response headers in
accordance with the SPDY 3 spec, and never produces further
response content. I've successfully run SGet against SServe to
get the 404, but with my current version of Chromium, SServe
reports that the TLS handshake fails. I believe this is because
my Chromium supports only SPDY 2, but I'll investigate further.

There are many TODOs left to deal with here, but I'll like to
get the skeleton of a server running for the happy cases before
handling all the error cases. The next step is to get SServe to
serve static files from a directory.