Serve static files from SServe.
Finally, the SServe program can serve static files from a given
directory. With this, I'm able to use SGet to download a copy of
/usr/share/dict/words, and the file transfers correctly. Asking for
a file that doesn't exist results in a 404 error, as it should.

This illustrates what it's like to implement a RequestHandler.
It's a little awkward, but this primitive interface will eventually
be hidden by something a little higher-level -- a WAI interface,
for example.

Doing this exercise has revealed one problem in Endpoint though:
a stream can terminate before all the WINDOW_UPDATE frames are
sent. For example, the default window size is 64 KB. The 'words'
file is a bit less than a megabyte, so SServe can go great guns
sending the first 64 KB worth of data frames. Shortly after the
last frame is sent, the code in Endpoint removes the stream, but
this happens before the last WINDOW_UPDATE frames arrive from
SGet. The incoming frame handler sends a RST_STREAM for each of
these WINDOW_UPDATE frames, which is not at all graceful.

Endpoint is arguably discarding the stream too early. Perhaps
the stream itself should have a callback that allows it to notify
the enclosing connection when no more frames are expected on the
stream, and we could delay removing the stream until all the
expected WINDOW_UPDATE frames have arrived.

Another problem is that a request handler still needs a way to
indicate that it has detected a protocol error, or suffered an
internal error. For example, if the push loop for a handler
encounters an IO exception reading from a static file, it should
be able to trigger a RST_STREAM with an internal error status.
Currently, we have no way to do this.

Despite the shortcomings, this is a nice little milestone, so in
it goes.