Separate the incoming and outgoing data window sizes.
For some reason, I had the impression that there was a single
data window size for both directions on a connection. However,
reading the spec more closely it seems that each side decides
on its window size for incoming data -- in fact, it may be
fixed in advance. This explains why the Google servers and
chromium send a SETTINGS frame with the window size when you
establish a connection.

Anyway, the point is that the incoming buffer size and the
outgoing data window size are in fact independent:

- the incoming buffer size -- which is the initial outgoing
  data window size to the remote endpoint -- is set to either
  the default (64KB) or a value communicated to the remote
  endpoint in a SETTINGS frame

- the outgoing window size is either the default (64KB) or
  whatever value is most recently received from the remote
  endpoint in a SETTINGS frame

In this commit, I've split out the two sizes in Stream and
Endpoint.Connection. I've also settled on the names

  IncomingBufferSize
  OutgoingWindowSize

which I think show the intent clearly.