Support exchanging window sizes with a remote endpoint.
Now Endpoint does two additional things:

1. on setting up a new connection, it immediately sends a
   SETTINGS frame with the connection's incoming buffer size
   as the data window size. This informs the remote endpoint
   of the incoming data window for this endpoint.

2. it handles incoming SETTINGS frames by extracting the
   window size setting, if present, and making that the
   intitial outgoing data window size for subsequent streams.

With these changes, I can now get chromium 21 (as of June 6)
to download a large file from sserve. The first time chromium
tries to connect, there is a handshake failure because of a
problem with the test certificate, but on the second attempt
it will work.

Chromium still never sends WINDOW_UPDATE frames, but it sets a
large data window size, so sserve doesn't exhaust the window
partway through the file. I'll have to check in the spdy-dev
list to see if there is any mention of whether this is a bug
in Chromium or if it's expected behaviour.

The interaction between sget and sserve still works, though I
haven't fixed the problem with trailing WINDOW_UPDATE frames
after a stream closes.