Don't call 'error' in BoundedBuffer.tryAdd for chunks bigger than the buffer.
The functions 'add' and 'tryAdd' would both invoke 'error' if the
given chunk was bigger than the total capacity of the buffer. For
the purposes of streams, it's more convenient if 'tryAdd' simply
returns 'False' for such a chunk. 'add' still invokes 'error',
since it cannot otherwise communicate the failure.

The bounded buffer unit tests did not fail when I made this change,
so there are obviously some coverage problems with the tests. I'll
fix that in a subsequent commit.