Add state snapshots to BoundedBuffer and Stream.
I've been trying to write some unit tests for Stream in one of
the styles described in the paper "Testing Monadic Code with
QuickCheck" (Coen Klassen and John Hughes). I've settled on the
abstract specification approach, where we model stream state
and operations on a pure data structure, and then show the
equivalence of the real and the pure versions. However, in order
to do that, we need functions to extract a snapshot of the
internal state of a Stream -- and also a BoundedBuffer.

I've added 'snapshot' functions to each module to do this. The
one for Stream is vulnerable to race conditions, since the
outgoing window size could change after we get it but before we
get the buffer snapshot, but this shouldn't be a problem for
the unit tests I plan to write.

The 'snapshot' function in BoundedBuffer makes redundant the
'remainingCapacity' function. However, this is used in the
existing unit tests for BoundedBuffer, so I'll remove it if and
when I rework the tests.