First crack at tests for Streams.
I finally got around to learning the approach to testing monadic
code intended by the Test.QuickCheck.Monadic module, and have
written a set of basic tests for Streams.

The approach here is described in the Klassen-Hughes paper
"Testing Monadic Code with QuickCheck", and uses one of the approaches
described there: testing that a Stream implements a specification
captured in a pure model, in randomly-generated reachable states.
To randomly generate a reachable state, we define a small language
of Stream programs, and provide a generator for legal Stream
programs. Here 'legal' means 'doesn't block indefinitely or throw
an exception.

There are still two things missing here:

- recording calls to the 'send window update' callback

- testing the blocking behaviour of the stream

However, it's taken me a while to get this far, and it's a good
checkpoint. I'll make the additions in subsequent commits.