Don't restrict the size of stream programs in StreamTests.
The 'forAllReachable' function used to restrict the size of stream
programs to 10 steps, since it greatly reduced the running time of
the stream tests. However, it turns out that the original performance
problems resulted from picking big initial window sizes, in the tens
or hundreds of kilobytes. By the time I made the commit that added
StreamTests.hs, I had reduced the window sizes in order to make the
error outputs easier to read -- it greatly shortened the generated
byte strings. I tried removing the upper limit on the program
length, and the tests are still quite fast.

I think that we're more likely to reach interesting stream states if
we allow the programs to grow as a test case progresses, so I've
removed the limit.