Add a couple of StreamModel tests, remove 'propM_refuseAddIncoming'.
I was thinking that one problem with 'propM_refuseAddIncoming' was
that while it checked that the model and the implementation agree,
it didn't check that both in fact refused to add the oversized
content. One way to deal with this is to write tests for the model
itself -- it's complicated enough to warrant it -- and then simply
allow both right-sized and oversized content in the content
generator for the 'propM_addIncoming' test case.

In other words, we confirm with ordinary QuickCheck tests that the
model behaves the way we want, and then confirm with the monadic
tests that the implementation corresponds to the model.

In the process, I've added the 'genReachableModel' generator.

At this point, I've added only tests for 'specAddIncoming'. I'll
add some more tests of the model shortly. For clarity, I've put
the model tests in their own separate test group, and run it before
the tests on the real implementation.