Upgrade the version of zlib-bindings.
This upgrades the version of zlib-bindings to 0.1.1.3, in order to
remove a dependency on a version of bytestring that precedes the one
in the 2013.2.0.0.

In this latest version of zlib-bindings, there is a Popper type
already defined as

 type Popper = IO (Maybe ByteString)

and the interface for dealing with compression and decompression are
changed and slightly different from one another. Because of the
differences, I've removed the generic 'transform' function and written
the 'compress' and 'decompress' functions directly.

According to the zlib-bindings Haddocks, 'flushInflate' is a synonym
for 'finishInflate'. It is not clear from the docs whether the Inflate
(decompression context) is still usable after calling
'flushInflate'. I'll find out shortly once I've upgraded
test-framework and test-framework-quickcheck and can run the tests.