Upgrade to the latest tls package and related packages.
Several old dependencies themselves have dependencies on packages that
are in the Haskell platform 2013.2.0.0, but on versions of those
packages that are older than the ones in the platform. Moreover, there
have been several updates and API changes in the past year, so it's
time to deal with upgrading.

The latest version of tls (1.1.5) and tls-extra (0.6.6) have some API
changes. In particular, there is now a single function for starting
either a TLS client or a TLS server; there are just differences in the
input parameters. This commit makes the necessary minor changes in
Network.SPDY.{Client,Server,NetworkConnection}.

Everything compiles, but I'm not yet sure that it runs.  I need to
make some other package upgrades before checking that.