Fix unit test warnings.
Compiling the test modules

 StreamTests
 BoundedBufferTests

would result in a warning

 Module `Prelude' does not export `catch'

It turns out that the Prelude module in base-4.6.0.0 and later no
longer exports the function

 catch :: IO a -> (IOError -> IO a) -> IO a

The two test modules had the import

 import Prelude hiding (catch)

In StreamTests, this wasn't needed. In BoundedBufferTests, it was
needed to avoid a name collision with Control.Exception.catch.

Since 'catch' was removed from the Prelude only in base-4.6, and I'd
rather have the minimum bound on base be 4, I've fixed this in
BoundedBufferTests by importing Control.Exception qualified to use
Control.Exception.catch, and then removed the explict Prelude import
from both test modules.