Move all the library sources below a 'src' directory.
In the interests of consistency, and because I'd also
like to expand the cabal file to build the example
programs, I've decided to move all the main library
source below a 'src' directory. Aside from the file
moves, this requires only adding an 'hs-source-dirs'
setting in the 'Library' section of the cabal file.

Everything still builds and the tests all pass.