Switch to at least version 1.0 of the 'options' package.
I realized that I had forgotten to check whether the example programs
compiled after upgrading to ghc 7.8.2. They didn't, because the latest
version of the 'options' package now has a different interface, new
since version 1.0.

The option declarations are a little longer now, since you have to
declare a data type for holding the option values, but this gets rid
of the dependency on Template Haskell, which is probably not such a
bad thing.

I've made the necessary changes in the five example programs, and set
the necessary lower bound on the depedencies for 'options'.